Fix uninstall (#462)
* Fix(core): secure / fix / enhanced uninstall process

* delete table for Model and Type

* fix MySQL query error: Identifier name  is too long

* adapt changelog.md

* fix CS

+++ inc/type.class.php
@@ -111,6 +111,26 @@ class PluginGenericobjectType extends CommonDBTM
         }
     }
 
+    public static function deleteFolder(string $folder)
+    {
+        if (!is_dir($folder)) {
+            return;
+        }
+
+        $it = new RecursiveDirectoryIterator($folder, FilesystemIterator::SKIP_DOTS);
+        $files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
+
+        foreach ($files as $file) {
+            if ($file->isDir()) {
+                rmdir($file->getRealPath());
+            } else {
+                unlink($file->getRealPath());
+            }
+        }
+
+        rmdir($folder);
+    }
+
 
     public static function getCompleteClassFilename($name)
     {
@@ -289,6 +309,16 @@ class PluginGenericobjectType extends CommonDBTM
         $preferences = new DisplayPreference();
         $preferences->deleteByCriteria(["itemtype" => $itemtype]);
         $DB->dropTable(getTableForItemType($itemtype), true);
+
+
+        if (strlen(getTableForItemType($itemtype) . "Model") <= 64) {
+            $DB->dropTable(getTableForItemType($itemtype . "Model"), true);
+        }
+
+        if (strlen(getTableForItemType($itemtype) . "Type") <= 64) {
+            $DB->dropTable(getTableForItemType($itemtype . "Type"), true);
+        }
+
     }
 
 
@@ -301,7 +331,9 @@ class PluginGenericobjectType extends CommonDBTM
     {
         /** @var DBmysql $DB */
         global $DB;
-        $DB->dropTable(getTableForItemType($itemtype) . "_items", true);
+        if (strlen(getTableForItemType($itemtype) . "_items") <= 64) {
+            $DB->dropTable(getTableForItemType($itemtype) . "_items", true);
+        }
     }
 
     /**
@@ -660,6 +692,13 @@ class PluginGenericobjectType extends CommonDBTM
         self::deleteItemtypeReferencesInGLPI(self::class);
 
         foreach ($DB->request(['FROM' => 'glpi_plugin_genericobject_types']) as $type) {
+
+            // GLPI 11 migration may change plugin itemtype from glpi_plugin_genericobject_types table during CustomAsset migration
+            // rely on original name to get correct itemtype
+            if (str_starts_with($type['itemtype'], 'Glpi\\CustomAsset\\')) {
+                $type['itemtype'] = self::getClassByName($type['name']);
+            }
+
             //Delete references to PluginGenericobjectType in the following tables
             self::deleteItemtypeReferencesInGLPI($type['itemtype']);
             //Dropd files and classes
@@ -668,6 +707,9 @@ class PluginGenericobjectType extends CommonDBTM
 
         //Delete table
         $migration->dropTable('glpi_plugin_genericobject_types');
+
+        self::deleteFolder(GENERICOBJECT_DOC_DIR);
+
     }
